ingest_bus/requirements_enhanced.txt
# Enhanced TORI Ingest Bus Requirements
# All file type support + system integrations

# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5
prometheus_client>=0.11.0

# Document processing dependencies
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.0.9
pandas>=1.3.0
mammoth>=1.4.0

# Text processing and NLP
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Optional ML dependencies for concept extraction
sentence-transformers>=2.0.0
transformers>=4.0.0

# Development and testing
pytest>=6.0.0
pytest-asyncio>=0.15.0
requests>=2.25.0

# System integration
aiohttp>=3.7.0
asyncpg>=0.24.0
redis>=3.5.0
